Allow the engine to read lines
