+++ src/engine.rs
@@ -1,3 +1,13 @@
+use std::io::{Stdout, Write};
+
+use crossterm::{
+    cursor::{position, MoveToColumn, RestorePosition, SavePosition},
+    event::{read, Event, KeyCode, KeyEvent, KeyModifiers},
+    style::{Color, Print, ResetColor, SetForegroundColor},
+    terminal::{Clear, ClearType},
+    ExecutableCommand, QueueableCommand, Result,
+};
+
 use std::collections::VecDeque;
 
 use crate::line_buffer::LineBuffer;
@@ -35,6 +45,42 @@ pub struct Engine {
     history: VecDeque<String>,
     history_cursor: i64,
     has_history: bool,
+}
+
+pub fn print_message(stdout: &mut Stdout, msg: &str) -> Result<()> {
+    stdout
+        .queue(Print("\n"))?
+        .queue(MoveToColumn(1))?
+        .queue(Print(msg))?
+        .queue(Print("\n"))?
+        .queue(MoveToColumn(1))?;
+    stdout.flush()?;
+
+    Ok(())
+}
+
+fn buffer_repaint(stdout: &mut Stdout, engine: &Engine, prompt_offset: u16) -> Result<()> {
+    let raw_buffer = engine.get_buffer();
+    let new_index = engine.get_insertion_point();
+
+    // Repaint logic:
+    //
+    // Start after the prompt
+    // Draw the string slice from 0 to the grapheme start left of insertion point
+    // Then, get the position on the screen
+    // Then draw the remainer of the buffer from above
+    // Finally, reset the cursor to the saved position
+
+    stdout.queue(MoveToColumn(prompt_offset))?;
+    stdout.queue(Print(&raw_buffer[0..new_index]))?;
+    stdout.queue(SavePosition)?;
+    stdout.queue(Print(&raw_buffer[new_index..]))?;
+    stdout.queue(Clear(ClearType::UntilNewLine))?;
+    stdout.queue(RestorePosition)?;
+
+    stdout.flush()?;
+
+    Ok(())
 }
 
 impl Engine {
@@ -256,5 +302,147 @@ impl Engine {
 
     pub fn move_word_right(&mut self) -> usize {
         self.line_buffer.move_word_right()
+    }
+
+    pub fn read_line(&mut self, stdout: &mut Stdout) -> Result<String> {
+        // print our prompt
+        stdout
+            .execute(SetForegroundColor(Color::Blue))?
+            .execute(Print("〉"))?
+            .execute(ResetColor)?;
+
+        // set where the input begins
+        let (mut prompt_offset, _) = position()?;
+        prompt_offset += 1;
+
+        loop {
+            match read()? {
+                Event::Key(KeyEvent {
+                    code,
+                    modifiers: KeyModifiers::CONTROL,
+                }) => match code {
+                    KeyCode::Char('d') => {
+                        if self.get_buffer().is_empty() {
+                            return Ok("exit".to_string());
+                        } else {
+                            self.run_edit_commands(&[EditCommand::Delete]);
+                        }
+                    }
+                    KeyCode::Char('a') => {
+                        self.run_edit_commands(&[EditCommand::MoveToStart]);
+                    }
+                    KeyCode::Char('e') => {
+                        self.run_edit_commands(&[EditCommand::MoveToEnd]);
+                    }
+                    KeyCode::Char('k') => {
+                        self.run_edit_commands(&[EditCommand::CutToEnd]);
+                    }
+                    KeyCode::Char('u') => {
+                        self.run_edit_commands(&[EditCommand::CutFromStart]);
+                    }
+                    KeyCode::Char('y') => {
+                        self.run_edit_commands(&[EditCommand::InsertCutBuffer]);
+                    }
+                    KeyCode::Char('b') => {
+                        self.run_edit_commands(&[EditCommand::MoveLeft]);
+                    }
+                    KeyCode::Char('f') => {
+                        self.run_edit_commands(&[EditCommand::MoveRight]);
+                    }
+                    KeyCode::Char('c') => {
+                        return Ok("".to_string());
+                    }
+                    KeyCode::Char('h') => {
+                        self.run_edit_commands(&[EditCommand::Backspace]);
+                    }
+                    KeyCode::Char('w') => {
+                        self.run_edit_commands(&[EditCommand::CutWordLeft]);
+                    }
+                    KeyCode::Left => {
+                        self.run_edit_commands(&[EditCommand::MoveWordLeft]);
+                    }
+                    KeyCode::Right => {
+                        self.run_edit_commands(&[EditCommand::MoveWordRight]);
+                    }
+                    _ => {}
+                },
+                Event::Key(KeyEvent {
+                    code,
+                    modifiers: KeyModifiers::ALT,
+                }) => match code {
+                    KeyCode::Char('b') => {
+                        self.run_edit_commands(&[EditCommand::MoveWordLeft]);
+                    }
+                    KeyCode::Char('f') => {
+                        self.run_edit_commands(&[EditCommand::MoveWordRight]);
+                    }
+                    KeyCode::Char('d') => {
+                        self.run_edit_commands(&[EditCommand::CutWordRight]);
+                    }
+                    KeyCode::Left => {
+                        self.run_edit_commands(&[EditCommand::MoveWordLeft]);
+                    }
+                    KeyCode::Right => {
+                        self.run_edit_commands(&[EditCommand::MoveWordRight]);
+                    }
+                    _ => {}
+                },
+                Event::Key(KeyEvent { code, modifiers: _ }) => {
+                    match code {
+                        KeyCode::Char(c) => {
+                            self.run_edit_commands(&[
+                                EditCommand::InsertChar(c),
+                                EditCommand::MoveRight,
+                            ]);
+                        }
+                        KeyCode::Backspace => {
+                            self.run_edit_commands(&[EditCommand::Backspace]);
+                        }
+                        KeyCode::Delete => {
+                            self.run_edit_commands(&[EditCommand::Delete]);
+                        }
+                        KeyCode::Home => {
+                            self.run_edit_commands(&[EditCommand::MoveToStart]);
+                        }
+                        KeyCode::End => {
+                            self.run_edit_commands(&[EditCommand::MoveToEnd]);
+                        }
+                        KeyCode::Enter => {
+                            let buffer = String::from(self.get_buffer());
+
+                            self.run_edit_commands(&[
+                                EditCommand::AppendToHistory,
+                                EditCommand::Clear,
+                            ]);
+
+                            return Ok(buffer);
+                        }
+                        KeyCode::Up => {
+                            self.run_edit_commands(&[EditCommand::PreviousHistory]);
+                        }
+                        KeyCode::Down => {
+                            // Down means: navigate forward through the history. If we reached the
+                            // bottom of the history, we clear the buffer, to make it feel like
+                            // zsh/bash/whatever
+                            self.run_edit_commands(&[EditCommand::NextHistory]);
+                        }
+                        KeyCode::Left => {
+                            self.run_edit_commands(&[EditCommand::MoveLeft]);
+                        }
+                        KeyCode::Right => {
+                            self.run_edit_commands(&[EditCommand::MoveRight]);
+                        }
+                        _ => {}
+                    };
+                }
+                Event::Mouse(event) => {
+                    print_message(stdout, &format!("{:?}", event))?;
+                }
+                Event::Resize(width, height) => {
+                    print_message(stdout, &format!("width: {} and height: {}", width, height))?;
+                }
+            }
+            buffer_repaint(stdout, &self, prompt_offset)?;
+        }
     }
 }
